move utility functions to a more sensible place for common usage
(with a slight refactor necessary for proper allround usage)
